refactor(client): route the socketless signal to the neutral diagnostic offer
Maintainer feedback on the first draft: VPN region-targeting is a CORE use of
BetterFleet, not an anomaly - the app must never tell players their VPN broke
detection, least of all on an unproven theory. The banner naming VPNs and ping
optimizers is gone, with its i18n block in all fourteen locales.

The valuable part stays: the Rust no_udp_cycles counter (game alive, UDP
enumeration empty cycle after cycle) still rides get_game_object, and the
watchdog - renamed InterceptorWatchdog -> SocketlessWatchdog, comments stripped
of any asserted cause - now raises the EXISTING #688 guided-diagnostic offer
instead of its own banner. Same neutral copy ('detection stayed silent - run a
capture?'), and the resulting report carries the evidence triage needs (empty
port lists, raw_packets) so the cause gets proven, not presumed.

Timing rules unchanged and still unit-tested: 3-minute sustained blind stretch,
3-cycle floor against enumeration hiccups, once per game-process lifetime,
countdown delays, never while detection works.
