+++ webapp/src-tauri/src/fetch_informations.rs
@@ -128,11 +128,11 @@ pub(crate) fn drop_quarantined(
 }
 
 /// Next value of the consecutive empty-UDP-enumeration counter (#801), the raw signal behind the
-/// frontend's "a VPN / ping optimizer intercepts the game's traffic" hint. Counts cycles where the
+/// frontend's socketless watchdog (which raises the neutral #688 diagnostic offer). Counts cycles where the
 /// game process exists but its UDP port enumeration came back empty; any cycle that finds ports
 /// (or the game closing) resets it. One or two empty cycles are ordinary enumeration hiccups
 /// (get_udp_connections returns empty on error too) - the field logs of #801 show the real
-/// interceptor signature as an empty enumeration EVERY cycle for many minutes, which is what a
+/// socketless signature as an empty enumeration EVERY cycle for many minutes, which is what a
 /// monotonically rising count encodes. Pure so the reset/increment rules are unit-tested.
 pub(crate) fn track_no_udp_cycles(prev: u32, game_running: bool, ports_empty: bool) -> u32 {
     if game_running && ports_empty {
@@ -222,7 +222,7 @@ pub async fn init() -> std::result::Result<Arc<RwLock<Api>>, anyhow::Error> {
             // grace decide, exactly as for a quiet capture window.
             let udp_ports = game_udp_candidate_ports(&game_pids);
             {
-                // Interceptor signal (#801): count consecutive empty enumerations while the game
+                // Socketless signal (report id 801): count consecutive empty enumerations while the game
                 // process exists; any cycle that finds ports resets it right below.
                 let mut api_lock = api.write().await;
                 api_lock.no_udp_cycles =
@@ -797,11 +797,11 @@ pub async fn create_raw_socket(socket_addr: SocketAddr) -> Result<UdpSocket> {
 mod tests {
     use super::*;
 
-    // The interceptor signal (#801): the counter must ride through the whole "game process alive,
+    // The socketless signal (report id 801): the counter must ride through the whole "game process alive,
     // enumeration empty" stretch and reset the moment either half stops holding.
     #[test]
     fn empty_enumerations_accumulate_only_while_the_game_runs() {
-        // A blind stretch: the count rises cycle after cycle - the VPN-interceptor signature.
+        // A blind stretch: the count rises cycle after cycle - the socketless signature.
         let mut cycles = 0;
         for _ in 0..3 {
             cycles = track_no_udp_cycles(cycles, true, true);
@@ -811,7 +811,7 @@ mod tests {
         assert_eq!(track_no_udp_cycles(cycles, true, false), 0);
         // Game gone: whatever was accumulated dies with the process.
         assert_eq!(track_no_udp_cycles(cycles, false, true), 0);
-        // Saturating, not wrapping: an interceptor left on for days must not reset the signal.
+        // Saturating, not wrapping: a blindness lasting days must not reset the signal.
         assert_eq!(track_no_udp_cycles(u32::MAX, true, true), u32::MAX);
     }
