+++ webapp/src/assets/locales/hi.json
@@ -206,10 +206,6 @@
       "title": "सर्वर पहचान निदान"
     },
     "dismiss": "अभी नहीं",
-    "interceptor": {
-      "banner": "ऐसा लगता है कि कोई VPN या पिंग ऑप्टिमाइज़र (ExitLag, WTFast…) गेम का ट्रैफ़िक इंटरसेप्ट कर रहा है — जब तक यह सक्रिय है, सर्वर पहचान काम नहीं कर सकती।",
-      "dismiss": "बंद करें"
-    },
     "prefill": "गेम में सर्वर पहचान शांत रही — स्वचालित निदान लॉग में संलग्न है।",
     "run": "निदान चलाएँ"
   },
