+++ webapp/src/assets/locales/ja.json
@@ -206,10 +206,6 @@
       "title": "サーバー検出の診断"
     },
     "dismiss": "今はしない",
-    "interceptor": {
-      "banner": "VPNやPing最適化ツール（ExitLag、WTFastなど）がゲームの通信を中継しているようです。有効になっている間、サーバー検出は機能しません。",
-      "dismiss": "閉じる"
-    },
     "prefill": "ゲーム中にサーバー検出が反応しませんでした。自動診断をログに添付しています。",
     "run": "診断を実行する"
   },
