+++ webapp/src/assets/locales/ko.json
@@ -206,10 +206,6 @@
       "title": "서버 감지 진단"
     },
     "dismiss": "나중에",
-    "interceptor": {
-      "banner": "VPN 또는 핑 최적화 프로그램(ExitLag, WTFast 등)이 게임 트래픽을 가로채고 있는 것 같습니다 — 활성화된 동안에는 서버 감지가 작동할 수 없습니다.",
-      "dismiss": "닫기"
-    },
     "prefill": "게임 내에서 서버 감지가 아무 반응이 없었습니다 — 자동 진단이 로그에 첨부되었습니다.",
     "run": "진단 실행"
   },
