+++ webapp/src/assets/locales/ru.json
@@ -206,10 +206,6 @@
       "title": "Диагностика определения сервера"
     },
     "dismiss": "Не сейчас",
-    "interceptor": {
-      "banner": "Похоже, VPN или оптимизатор пинга (ExitLag, WTFast…) перехватывает трафик игры — пока он активен, определение сервера работать не может.",
-      "dismiss": "Закрыть"
-    },
     "prefill": "Определение сервера молчало в игре — автоматическая диагностика приложена в логах.",
     "run": "Запустить диагностику"
   },
