+++ webapp/src/assets/locales/uk.json
@@ -206,10 +206,6 @@
       "title": "Діагностика виявлення сервера"
     },
     "dismiss": "Не зараз",
-    "interceptor": {
-      "banner": "Схоже, VPN або оптимізатор пінгу (ExitLag, WTFast…) перехоплює трафік гри — поки він активний, виявлення сервера не може працювати.",
-      "dismiss": "Закрити"
-    },
     "prefill": "Виявлення сервера мовчало у грі — автоматичну діагностику додано до логів.",
     "run": "Запустити діагностику"
   },
