+++ webapp/src/assets/locales/zh.json
@@ -206,10 +206,6 @@
       "title": "服务器检测诊断"
     },
     "dismiss": "暂不",
-    "interceptor": {
-      "banner": "VPN 或延迟优化工具（ExitLag、WTFast 等）似乎正在拦截游戏流量——在其开启期间，服务器检测无法工作。",
-      "dismiss": "关闭"
-    },
     "prefill": "服务器检测在游戏中始终没有响应——已在日志中附上自动诊断信息。",
     "run": "运行诊断"
   },
